+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - added Copy File Function
 - add Built for Linux aarch 64
 - added dotenv parsing
+- added Download Shell Script
 
 ## 0.1.10 - 05.10.2025
 - added Function to Check the OS Type to make code in the script which runs only